+++ src/__internal__/input/index.d.ts
@@ -1,5 +1,0 @@
-export { default, InputProps, CommonInputProps } from "./input";
-export {
-  InputPresentationProps,
-  CommonInputPresentationProps,
-} from "./input-presentation";